feat(streaming): provide `afterGoLive` hook
Provide a hook that could be invoked by platform services,
after finishing going live. We're notified of live status
through signal handling, we chose to invoke such hook
directly if exists, as opposed to letting the client subscribe,
since there are many layers of indirection.

`startStreaming` and `toggleStreaming` now accept an optional
context, which is passed to the aforementioned `afterGoLive`
hook. This is necessary if the caller wants to use some data
that was passed prior to the stream starting (for example:
tags are passed through the `EditStreamInfo` component prior
to the stream being live, Twitch requires that the stream is
live in order to update tags, so we pass the tags before starting
and receive them on the hook.